requirements.txt
torch==2.2.0 torchvision==0.17.0 torchaudio==2.2.0 --index-url https://download.pytorch.org/whl/cu118
datasets
evaluate
transformers==4.46.2
scikit-learn
sentencepiece
protobuf
accelerate
jsonlines
pillow
peft
timm==0.9.16
torch-pruning==1.3.7
scypy
huggingface-hub
openai
anthropic
google-genai
jupyterlab
notebook
pandas
wheel
opencv-python-headless
scikit-image